online_exam_cheating_detection/requirements.txt
Flask
tensorflow-cpu
opencv-python
